refactor: Rename the package to discover_test_coverage.
