+++ discover_test_coverage/server.py
@@ -1,7 +1,7 @@
 """Create and run a syslog remote server."""
 
-from discover import constants
-from discover import output
+from discover_test_coverage import constants
+from discover_test_coverage import output
 
 import logging
 import logging.handlers
